♻️ Extracted Sentry and FirstPromoter setup to setup-integrations.js
Extracted two initialization functions to improve code organization:
- setupSentry() - Initialize Sentry error tracking
- setupFirstPromoter() - Load and initialize FirstPromoter tracking script

Reduced App.js by 74 lines (802 → 728 lines)
All tests passing (256/257)
